+++ log/table.go
@@ -3,13 +3,14 @@ package log
 import (
 	"time"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
 // Table makes trace.Table with internal logging
 // nolint:gocyclo
-func Table(log Logger, details trace.Details) (t trace.Table) {
-	log = log.WithName(`table`)
+func Table(l Logger, details trace.Details) (t trace.Table) {
+	l = l.WithName(`table`)
 	// nolint:nestif
 	if details&trace.TableEvents != 0 {
 		t.OnDo = func(
@@ -20,35 +21,52 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 			trace.TableDoDoneInfo,
 		) {
 			idempotent := info.Idempotent
-			log.Tracef(`do start {idempotent:%t}`,
+			l.Tracef(`do start {idempotent:%t}`,
 				idempotent,
 			)
 			start := time.Now()
 			return func(info trace.TableDoIntermediateInfo) func(trace.TableDoDoneInfo) {
 				if info.Error == nil {
-					log.Tracef(`do intermediate {latency:"%v",idempotent:%t}`,
+					l.Tracef(`do attempt done {latency:"%v",idempotent:%t}`,
 						time.Since(start),
 						idempotent,
 					)
 				} else {
-					log.Debugf(`do intermediate {latency:"%v",idempotent:%t,error:"%v"}`,
+					log := l.Warnf
+					m := retry.Check(info.Error)
+					if m.StatusCode() < 0 {
+						log = l.Debugf
+					}
+					log(`do attempt failed {latency:"%v",idempotent:%t,error:"%s",retryable:%t,code:%d,deleteSession:%t}`,
 						time.Since(start),
 						idempotent,
 						info.Error,
+						m.MustRetry(idempotent),
+						m.StatusCode(),
+						m.MustDeleteSession(),
 					)
 				}
 				return func(info trace.TableDoDoneInfo) {
 					if info.Error == nil {
-						log.Tracef(`do done {latency:"%v",idempotent:%t,attempts:%d}`,
+						l.Tracef(`do done {latency:"%v",idempotent:%t,attempts:%d}`,
 							time.Since(start),
 							idempotent,
 							info.Attempts,
 						)
 					} else {
-						log.Errorf(`do intermediate {latency:"%v",idempotent:%t,error:"%v"}`,
+						log := l.Errorf
+						m := retry.Check(info.Error)
+						if m.StatusCode() < 0 {
+							log = l.Debugf
+						}
+						log(`do failed {latency:"%v",idempotent:%t,attempts:%d,error:"%s",retryable:%t,code:%d,deleteSession:%t}`,
 							time.Since(start),
 							idempotent,
+							info.Attempts,
 							info.Error,
+							m.MustRetry(idempotent),
+							m.StatusCode(),
+							m.MustDeleteSession(),
 						)
 					}
 				}
@@ -62,36 +80,52 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 			trace.TableDoTxDoneInfo,
 		) {
 			idempotent := info.Idempotent
-			log.Tracef(`doTx start {idempotent:%t}`,
+			l.Tracef(`doTx start {idempotent:%t}`,
 				idempotent,
 			)
 			start := time.Now()
 			return func(info trace.TableDoTxIntermediateInfo) func(trace.TableDoTxDoneInfo) {
 				if info.Error == nil {
-					log.Tracef(`doTx intermediate {latency:"%v",idempotent:%t}`,
+					l.Tracef(`doTx attempt done {latency:"%v",idempotent:%t}`,
 						time.Since(start),
 						idempotent,
 					)
 				} else {
-					log.Errorf(`doTx intermediate {latency:"%v",idempotent:%t,error:"%v"}`,
+					log := l.Warnf
+					m := retry.Check(info.Error)
+					if m.StatusCode() < 0 {
+						log = l.Debugf
+					}
+					log(`doTx attempt failed {latency:"%v",idempotent:%t,error:"%s",retryable:%t,code:%d,deleteSession:%t}`,
 						time.Since(start),
 						idempotent,
 						info.Error,
+						m.MustRetry(idempotent),
+						m.StatusCode(),
+						m.MustDeleteSession(),
 					)
 				}
 				return func(info trace.TableDoTxDoneInfo) {
 					if info.Error == nil {
-						log.Tracef(`doTx done {latency:"%v",idempotent:%t,attempts:%d}`,
+						l.Tracef(`doTx done {latency:"%v",idempotent:%t,attempts:%d}`,
 							time.Since(start),
 							idempotent,
 							info.Attempts,
 						)
 					} else {
-						log.Errorf(`doTx failed {latency:"%v",idempotent:%t,attempts:%d,error:"%v"}`,
+						log := l.Errorf
+						m := retry.Check(info.Error)
+						if m.StatusCode() < 0 {
+							log = l.Debugf
+						}
+						log(`doTx failed {latency:"%v",idempotent:%t,attempts:%d,error:"%s",retryable:%t,code:%d,deleteSession:%t}`,
 							time.Since(start),
 							idempotent,
 							info.Attempts,
 							info.Error,
+							m.MustRetry(idempotent),
+							m.StatusCode(),
+							m.MustDeleteSession(),
 						)
 					}
 				}
@@ -104,29 +138,29 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 		) func(
 			trace.TableCreateSessionDoneInfo,
 		) {
-			log.Tracef(`create session start`)
+			l.Tracef(`create session start`)
 			start := time.Now()
 			return func(info trace.TableCreateSessionIntermediateInfo) func(trace.TableCreateSessionDoneInfo) {
 				if info.Error == nil {
-					log.Tracef(`create session intermediate {latency:"%v"}`,
+					l.Tracef(`create session intermediate {latency:"%v"}`,
 						time.Since(start),
 					)
 				} else {
-					log.Errorf(`create session intermediate {latency:"%v",error:"%v"}`,
+					l.Errorf(`create session intermediate {latency:"%v",error:"%v"}`,
 						time.Since(start),
 						info.Error,
 					)
 				}
 				return func(info trace.TableCreateSessionDoneInfo) {
 					if info.Error == nil {
-						log.Tracef(`create session done {latency:"%v",attempts:%d,session:{id:"%s",status:"%s"}}`,
+						l.Tracef(`create session done {latency:"%v",attempts:%d,session:{id:"%s",status:"%s"}}`,
 							time.Since(start),
 							info.Attempts,
 							info.Session.ID(),
 							info.Session.Status(),
 						)
 					} else {
-						log.Errorf(`create session failed {latency:"%v",attempts:%d,error:"%v"}`,
+						l.Errorf(`create session failed {latency:"%v",attempts:%d,error:"%v"}`,
 							time.Since(start),
 							info.Attempts,
 							info.Error,
@@ -139,25 +173,25 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 	// nolint:nestif
 	if details&trace.TableSessionEvents != 0 {
 		// nolint:govet
-		log := log.WithName(`session`)
+		l := l.WithName(`session`)
 		if details&trace.TableSessionLifeCycleEvents != 0 {
 			t.OnSessionNew = func(info trace.TableSessionNewStartInfo) func(trace.TableSessionNewDoneInfo) {
-				log.Tracef(`create start`)
+				l.Tracef(`create start`)
 				start := time.Now()
 				return func(info trace.TableSessionNewDoneInfo) {
 					if info.Error == nil {
 						if info.Session != nil {
-							log.Tracef(`create done {latency:"%v",id:%d}`,
+							l.Tracef(`create done {latency:"%v",id:%d}`,
 								time.Since(start),
 								info.Session.ID(),
 							)
 						} else {
-							log.Warnf(`create done without session {latency:"%v"}`,
+							l.Warnf(`create done without session {latency:"%v"}`,
 								time.Since(start),
 							)
 						}
 					} else {
-						log.Warnf(`create failed {latency:"%v",error:"%v"}`,
+						l.Warnf(`create failed {latency:"%v",error:"%v"}`,
 							time.Since(start),
 							info.Error,
 						)
@@ -166,20 +200,20 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 			}
 			t.OnSessionDelete = func(info trace.TableSessionDeleteStartInfo) func(trace.TableSessionDeleteDoneInfo) {
 				session := info.Session
-				log.Tracef(`delete start {id:"%s",status:"%s"}`,
+				l.Tracef(`delete start {id:"%s",status:"%s"}`,
 					session.ID(),
 					session.Status(),
 				)
 				start := time.Now()
 				return func(info trace.TableSessionDeleteDoneInfo) {
 					if info.Error == nil {
-						log.Tracef(`delete done {latency:"%v",id:"%s",status:"%s"}`,
+						l.Tracef(`delete done {latency:"%v",id:"%s",status:"%s"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
 						)
 					} else {
-						log.Warnf(`delete failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
+						l.Warnf(`delete failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
@@ -190,20 +224,20 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 			}
 			t.OnSessionKeepAlive = func(info trace.TableKeepAliveStartInfo) func(trace.TableKeepAliveDoneInfo) {
 				session := info.Session
-				log.Tracef(`keep-alive start {id:"%s",status:"%s"}`,
+				l.Tracef(`keep-alive start {id:"%s",status:"%s"}`,
 					session.ID(),
 					session.Status(),
 				)
 				start := time.Now()
 				return func(info trace.TableKeepAliveDoneInfo) {
 					if info.Error == nil {
-						log.Tracef(`keep-alive done {latency:"%v",id:"%s",status:"%s"}`,
+						l.Tracef(`keep-alive done {latency:"%v",id:"%s",status:"%s"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
 						)
 					} else {
-						log.Warnf(`keep-alive failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
+						l.Warnf(`keep-alive failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
@@ -215,10 +249,10 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 		}
 		if details&trace.TableSessionQueryEvents != 0 {
 			// nolint:govet
-			log := log.WithName(`query`)
+			l := l.WithName(`query`)
 			if details&trace.TableSessionQueryInvokeEvents != 0 {
 				// nolint:govet
-				log := log.WithName(`invoke`)
+				l := l.WithName(`invoke`)
 				t.OnSessionQueryPrepare = func(
 					info trace.TablePrepareDataQueryStartInfo,
 				) func(
@@ -226,7 +260,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 				) {
 					session := info.Session
 					query := info.Query
-					log.Tracef(`prepare start {id:"%s",status:"%s",query:"%s"}`,
+					l.Tracef(`prepare start {id:"%s",status:"%s",query:"%s"}`,
 						session.ID(),
 						session.Status(),
 						query,
@@ -234,7 +268,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 					start := time.Now()
 					return func(info trace.TablePrepareDataQueryDoneInfo) {
 						if info.Error == nil {
-							log.Debugf(`prepare done {latency:"%v",id:"%s",status:"%s",query:"%s",result:"%s"}`,
+							l.Debugf(`prepare done {latency:"%v",id:"%s",status:"%s",query:"%s",result:"%s"}`,
 								time.Since(start),
 								session.ID(),
 								session.Status(),
@@ -242,7 +276,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 								info.Result,
 							)
 						} else {
-							log.Errorf(`prepare failed {latency:"%v",id:"%s",status:"%s",query:"%s",error:"%v"}`,
+							l.Errorf(`prepare failed {latency:"%v",id:"%s",status:"%s",query:"%s",error:"%v"}`,
 								time.Since(start),
 								session.ID(),
 								session.Status(),
@@ -260,7 +294,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 					session := info.Session
 					query := info.Query
 					params := info.Parameters
-					log.Tracef(`execute start {id:"%s",status:"%s",query:"%s",params:"%s"}`,
+					l.Tracef(`execute start {id:"%s",status:"%s",query:"%s",params:"%s"}`,
 						session.ID(),
 						session.Status(),
 						query,
@@ -270,7 +304,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 					return func(info trace.TableExecuteDataQueryDoneInfo) {
 						if info.Error == nil {
 							tx := info.Tx
-							log.Debugf(
+							l.Debugf(
 								`execute done {latency:"%v",id:"%s",status:"%s",tx:"%s",query:"%s",params:"%s",prepared:%t,result:{err:"%v"}}`,
 								time.Since(start),
 								session.ID(),
@@ -282,7 +316,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 								info.Result.Err(),
 							)
 						} else {
-							log.Errorf(`execute failed {latency:"%v",id:"%s",status:"%s",query:"%s",params:"%s",prepared:%t,error:"%v"}`,
+							l.Errorf(`execute failed {latency:"%v",id:"%s",status:"%s",query:"%s",params:"%s",prepared:%t,error:"%v"}`,
 								time.Since(start),
 								session.ID(),
 								session.Status(),
@@ -297,7 +331,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 			}
 			if details&trace.TableSessionQueryStreamEvents != 0 {
 				// nolint:govet
-				log := log.WithName(`stream`)
+				l := l.WithName(`stream`)
 				t.OnSessionQueryStreamExecute = func(
 					info trace.TableSessionQueryStreamExecuteStartInfo,
 				) func(
@@ -308,7 +342,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 					session := info.Session
 					query := info.Query
 					params := info.Parameters
-					log.Tracef(`stream execute start {id:"%s",status:"%s",query:"%s",params:"%s"}`,
+					l.Tracef(`stream execute start {id:"%s",status:"%s",query:"%s",params:"%s"}`,
 						session.ID(),
 						session.Status(),
 						query,
@@ -321,15 +355,15 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 						trace.TableSessionQueryStreamExecuteDoneInfo,
 					) {
 						if info.Error == nil {
-							log.Tracef(`stream execute intermediate`)
+							l.Tracef(`stream execute intermediate`)
 						} else {
-							log.Warnf(`stream execute intermediate failed {error:"%v"}`,
+							l.Warnf(`stream execute intermediate failed {error:"%v"}`,
 								info.Error,
 							)
 						}
 						return func(info trace.TableSessionQueryStreamExecuteDoneInfo) {
 							if info.Error == nil {
-								log.Debugf(`stream execute done {latency:"%v",id:"%s",status:"%s",query:"%s",params:"%s"}`,
+								l.Debugf(`stream execute done {latency:"%v",id:"%s",status:"%s",query:"%s",params:"%s"}`,
 									time.Since(start),
 									session.ID(),
 									session.Status(),
@@ -337,7 +371,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 									params,
 								)
 							} else {
-								log.Errorf(`stream execute failed {latency:"%v",id:"%s",status:"%s",query:"%s",params:"%s",error:"%v"}`,
+								l.Errorf(`stream execute failed {latency:"%v",id:"%s",status:"%s",query:"%s",params:"%s",error:"%v"}`,
 									time.Since(start),
 									session.ID(),
 									session.Status(),
@@ -357,7 +391,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 					trace.TableSessionQueryStreamReadDoneInfo,
 				) {
 					session := info.Session
-					log.Tracef(`read start {id:"%s",status:"%s"}`,
+					l.Tracef(`read start {id:"%s",status:"%s"}`,
 						session.ID(),
 						session.Status(),
 					)
@@ -368,21 +402,21 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 						trace.TableSessionQueryStreamReadDoneInfo,
 					) {
 						if info.Error == nil {
-							log.Tracef(`intermediate`)
+							l.Tracef(`intermediate`)
 						} else {
-							log.Warnf(`intermediate failed {error:"%v"}`,
+							l.Warnf(`intermediate failed {error:"%v"}`,
 								info.Error,
 							)
 						}
 						return func(info trace.TableSessionQueryStreamReadDoneInfo) {
 							if info.Error == nil {
-								log.Debugf(`read done {latency:"%v",id:"%s",status:"%s"}`,
+								l.Debugf(`read done {latency:"%v",id:"%s",status:"%s"}`,
 									time.Since(start),
 									session.ID(),
 									session.Status(),
 								)
 							} else {
-								log.Errorf(`read failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
+								l.Errorf(`read failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
 									time.Since(start),
 									session.ID(),
 									session.Status(),
@@ -396,28 +430,28 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 		}
 		if details&trace.TableSessionTransactionEvents != 0 {
 			// nolint:govet
-			log := log.WithName(`transaction`)
+			l := l.WithName(`transaction`)
 			t.OnSessionTransactionBegin = func(
 				info trace.TableSessionTransactionBeginStartInfo,
 			) func(
 				trace.TableSessionTransactionBeginDoneInfo,
 			) {
 				session := info.Session
-				log.Tracef(`begin start {id:"%s",status:"%s"}`,
+				l.Tracef(`begin start {id:"%s",status:"%s"}`,
 					session.ID(),
 					session.Status(),
 				)
 				start := time.Now()
 				return func(info trace.TableSessionTransactionBeginDoneInfo) {
 					if info.Error == nil {
-						log.Debugf(`begin done {latency:"%v",id:"%s",status:"%s",tx:"%s"}`,
+						l.Debugf(`begin done {latency:"%v",id:"%s",status:"%s",tx:"%s"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
 							info.Tx.ID(),
 						)
 					} else {
-						log.Warnf(`begin failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
+						l.Warnf(`begin failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
@@ -433,7 +467,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 			) {
 				session := info.Session
 				tx := info.Tx
-				log.Tracef(`commit start {id:"%s",status:"%s",tx:"%s"}`,
+				l.Tracef(`commit start {id:"%s",status:"%s",tx:"%s"}`,
 					session.ID(),
 					session.Status(),
 					tx.ID(),
@@ -441,14 +475,14 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 				start := time.Now()
 				return func(info trace.TableSessionTransactionCommitDoneInfo) {
 					if info.Error == nil {
-						log.Debugf(`commit done {latency:"%v",id:"%s",status:"%s",tx:"%s"}`,
+						l.Debugf(`commit done {latency:"%v",id:"%s",status:"%s",tx:"%s"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
 							tx.ID(),
 						)
 					} else {
-						log.Errorf(`commit failed {latency:"%v",id:"%s",status:"%s",tx:"%s",error:"%v"}`,
+						l.Errorf(`commit failed {latency:"%v",id:"%s",status:"%s",tx:"%s",error:"%v"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
@@ -465,7 +499,7 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 			) {
 				session := info.Session
 				tx := info.Tx
-				log.Tracef(`rollback start {id:"%s",status:"%s",tx:"%s"}`,
+				l.Tracef(`rollback start {id:"%s",status:"%s",tx:"%s"}`,
 					session.ID(),
 					session.Status(),
 					tx.ID(),
@@ -473,14 +507,14 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 				start := time.Now()
 				return func(info trace.TableSessionTransactionRollbackDoneInfo) {
 					if info.Error == nil {
-						log.Debugf(`rollback done {latency:"%v",id:"%s",status:"%s",tx:"%s"}`,
+						l.Debugf(`rollback done {latency:"%v",id:"%s",status:"%s",tx:"%s"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
 							tx.ID(),
 						)
 					} else {
-						log.Errorf(`rollback failed {latency:"%v",id:"%s",status:"%s",tx:"%s",error:"%v"}`,
+						l.Errorf(`rollback failed {latency:"%v",id:"%s",status:"%s",tx:"%s",error:"%v"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
@@ -495,13 +529,13 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 	// nolint:nestif
 	if details&trace.TablePoolEvents != 0 {
 		// nolint:govet
-		log := log.WithName(`pool`)
+		l := l.WithName(`pool`)
 		if details&trace.TablePoolLifeCycleEvents != 0 {
 			t.OnInit = func(info trace.TableInitStartInfo) func(trace.TableInitDoneInfo) {
-				log.Infof(`initialize start`)
+				l.Infof(`initialize start`)
 				start := time.Now()
 				return func(info trace.TableInitDoneInfo) {
-					log.Infof(`initialize done {latency:"%v",size:{min:%d,max:%d}}`,
+					l.Infof(`initialize done {latency:"%v",size:{min:%d,max:%d}}`,
 						time.Since(start),
 						info.KeepAliveMinSize,
 						info.Limit,
@@ -509,15 +543,15 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 				}
 			}
 			t.OnClose = func(info trace.TableCloseStartInfo) func(trace.TableCloseDoneInfo) {
-				log.Infof(`close start`)
+				l.Infof(`close start`)
 				start := time.Now()
 				return func(info trace.TableCloseDoneInfo) {
 					if info.Error == nil {
-						log.Infof(`close done {latency:"%v"}`,
+						l.Infof(`close done {latency:"%v"}`,
 							time.Since(start),
 						)
 					} else {
-						log.Errorf(`close failed {latency:"%v",error:"%v"}`,
+						l.Errorf(`close failed {latency:"%v",error:"%v"}`,
 							time.Since(start),
 							info.Error,
 						)
@@ -525,27 +559,27 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 				}
 			}
 			t.OnPoolStateChange = func(info trace.TablePooStateChangeInfo) {
-				log.Infof(`state changed {size:%d,event:"%s"}`, info.Size, info.Event)
+				l.Infof(`state changed {size:%d,event:"%s"}`, info.Size, info.Event)
 			}
 		}
 		if details&trace.TablePoolSessionLifeCycleEvents != 0 {
 			// nolint:govet
-			log := log.WithName(`session`)
+			l := l.WithName(`session`)
 			t.OnPoolSessionNew = func(info trace.TablePoolSessionNewStartInfo) func(trace.TablePoolSessionNewDoneInfo) {
-				log.Tracef(`create start`)
+				l.Tracef(`create start`)
 				start := time.Now()
 				return func(info trace.TablePoolSessionNewDoneInfo) {
 					if info.Error == nil {
 						session := info.Session
 						if session != nil {
-							log.Debugf(`create done {latency:"%v",id:"%s",status:"%s"}`,
+							l.Debugf(`create done {latency:"%v",id:"%s",status:"%s"}`,
 								time.Since(start),
 								session.ID(),
 								session.Status(),
 							)
 						}
 					} else {
-						log.Errorf(`create failed {latency:"%v",error:"%v"}`,
+						l.Errorf(`create failed {latency:"%v",error:"%v"}`,
 							time.Since(start),
 							info.Error,
 						)
@@ -554,13 +588,13 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 			}
 			t.OnPoolSessionClose = func(info trace.TablePoolSessionCloseStartInfo) func(trace.TablePoolSessionCloseDoneInfo) {
 				session := info.Session
-				log.Tracef(`close start {id:"%s",status:"%s"}`,
+				l.Tracef(`close start {id:"%s",status:"%s"}`,
 					session.ID(),
 					session.Status(),
 				)
 				start := time.Now()
 				return func(info trace.TablePoolSessionCloseDoneInfo) {
-					log.Debugf(`close done {latency:"%v",id:"%s",status:"%s"}`,
+					l.Debugf(`close done {latency:"%v",id:"%s",status:"%s"}`,
 						time.Since(start),
 						session.ID(),
 						session.Status(),
@@ -571,20 +605,20 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 		if details&trace.TablePoolAPIEvents != 0 {
 			t.OnPoolPut = func(info trace.TablePoolPutStartInfo) func(trace.TablePoolPutDoneInfo) {
 				session := info.Session
-				log.Tracef(`put start {id:"%s",status:"%s"}`,
+				l.Tracef(`put start {id:"%s",status:"%s"}`,
 					session.ID(),
 					session.Status(),
 				)
 				start := time.Now()
 				return func(info trace.TablePoolPutDoneInfo) {
 					if info.Error == nil {
-						log.Tracef(`put done {latency:"%v",id:"%s",status:"%s"}`,
+						l.Tracef(`put done {latency:"%v",id:"%s",status:"%s"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
 						)
 					} else {
-						log.Errorf(`put failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
+						l.Errorf(`put failed {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
@@ -594,19 +628,19 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 				}
 			}
 			t.OnPoolGet = func(info trace.TablePoolGetStartInfo) func(trace.TablePoolGetDoneInfo) {
-				log.Tracef(`get start`)
+				l.Tracef(`get start`)
 				start := time.Now()
 				return func(info trace.TablePoolGetDoneInfo) {
 					if info.Error == nil {
 						session := info.Session
-						log.Tracef(`get done {latency:"%v",id:"%s",status:"%s",attempts:%d}`,
+						l.Tracef(`get done {latency:"%v",id:"%s",status:"%s",attempts:%d}`,
 							time.Since(start),
 							session.ID(),
 							session.Status(),
 							info.Attempts,
 						)
 					} else {
-						log.Warnf(`get failed {latency:"%v",attempts:%d,error:"%v"}`,
+						l.Warnf(`get failed {latency:"%v",attempts:%d,error:"%v"}`,
 							time.Since(start),
 							info.Attempts,
 							info.Error,
@@ -615,16 +649,16 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 				}
 			}
 			t.OnPoolWait = func(info trace.TablePoolWaitStartInfo) func(trace.TablePoolWaitDoneInfo) {
-				log.Tracef(`wait start`)
+				l.Tracef(`wait start`)
 				start := time.Now()
 				return func(info trace.TablePoolWaitDoneInfo) {
 					if info.Error == nil {
 						if info.Session == nil {
-							log.Tracef(`wait done without any significant result {latency:"%v"}`,
+							l.Tracef(`wait done without any significant result {latency:"%v"}`,
 								time.Since(start),
 							)
 						} else {
-							log.Tracef(`wait done {latency:"%v",id:"%s",status:"%s"}`,
+							l.Tracef(`wait done {latency:"%v",id:"%s",status:"%s"}`,
 								time.Since(start),
 								info.Session.ID(),
 								info.Session.Status(),
@@ -632,12 +666,12 @@ func Table(log Logger, details trace.Details) (t trace.Table) {
 						}
 					} else {
 						if info.Session == nil {
-							log.Tracef(`wait failed {latency:"%v",error:"%v"}`,
+							l.Tracef(`wait failed {latency:"%v",error:"%v"}`,
 								time.Since(start),
 								info.Error,
 							)
 						} else {
-							log.Tracef(`wait failed with session {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
+							l.Tracef(`wait failed with session {latency:"%v",id:"%s",status:"%s",error:"%v"}`,
 								time.Since(start),
 								info.Session.ID(),
 								info.Session.Status(),